hardware-integration/rfid/requirements.txt
# RFID Reader Requirements
RPi.GPIO==0.7.1
mfrc522==0.0.7
requests==2.31.0
